First version of the ArticleAPI
First version of the ArticleAPI, used by the Article and ArticleVersion microservice.
For now I have specified only four methods GET, I won't detail the query parameters or the responses, those are in the documentation but they methods are:
-GET /article/{id} -> get an Article by ID
-GET /article/{name} -> get an Article by name (in a given wiki, but not yet implemented)
-GET /articleVersion/{id} -> get an ArticleVersion by ID
-GET /article/search -> searches for articles in a given wiki with certain parameters. Designed to ease pagination.

Work to do: Implement more useful and necesary GET methods as well as PUT, DELETE, etc. methods.