ThemeContext created
Defined a context that provides the current theme and the function to toggle it.

It has been created using the Ant Design themes algorithm. The algorithm provides a way to derivate some secondary colors by defining some primary colors as a base.

We are using the light and dark default preset that Ant design provides.

You can use the css variables it generates on your css files