Header and profile popover translations
Created a bunch of new files that manage the localization.

- i18n.js : this is the root configuration file for the i18next package that we are using to manage the translation files. It specifies that the default language is english and the trans folders.

- header-trans.json: json file that that contain key:value pairs. Each value represent the translation of that key in each language.

For example, when the current language is "es" (i.e Spanish), the i18n package will look into the es/ header-trans file to seach the key-value pairs.
To add the translations for every context of the application, we just need to add a new file