APIGateway - security method implemented
**NOT TESTED**

Implemented the cache with the redis libray, which basically creates a local database. I storage (token, token_info) key-value pairs with a TTL of 1.5 hours.
That means that expired tokens will be erased over time.

Regurlarly, tokens last for 1 hour, so I made that they are stored for a little longer because, checking its expiration date is faster on the gateway than on the users microservice (because it has to check it with google).